add Swipe Screen

+++ composeApp/src/commonMain/kotlin/com/github/freshmorsikov/moviematcher/util/NumberUtil.kt
@@ -1,5 +1,5 @@
 package com.github.freshmorsikov.moviematcher.util
 
-fun Float.toRatingFormat(): String {
+fun Double.toRatingFormat(): String {
     return this.toString().take(3)
 }
